Initial attempt at using UpdateKernel for the updatedb command.
