+++ src/Boot/DrupalBoot8.php
@@ -9,7 +9,7 @@ use Drupal\Core\DrupalKernel;
 use Drush\Drush;
 use Drush\Drupal\DrupalKernel as DrushDrupalKernel;
 use Drush\Drupal\DrushServiceModifier;
-use Drupal\Core\Database\Database;
+use Drush\Drupal\UpdateKernel as DrushUpdateKernel;
 
 use Drush\Log\LogLevel;
 
@@ -139,12 +139,13 @@ class DrupalBoot8 extends DrupalBoot implements AutoloaderAwareInterface
         parent::bootstrapDrupalDatabase();
     }
 
-    public function bootstrapDrupalConfiguration()
+    public function bootstrapDrupalConfiguration($kernel = 'drupal')
     {
         $classloader = $this->autoloader();
-        $kernelClass = new \ReflectionClass('\Drupal\Core\DrupalKernel');
         $request = $this->getRequest();
-        $this->kernel = DrushDrupalKernel::createFromRequest($request, $classloader, 'prod');
+        $kernel_factory = $this->getKernelFactory($kernel);
+        /** @var \Drupal\Core\DrupalKernelInterface kernel */
+        $this->kernel = $kernel_factory($request, $classloader, 'prod');
         // @see Drush\Drupal\DrupalKernel::addServiceModifier()
         $this->kernel->addServiceModifier(new DrushServiceModifier());
 
@@ -223,5 +224,23 @@ class DrupalBoot8 extends DrupalBoot implements AutoloaderAwareInterface
             $response = Response::create('');
             $this->kernel->terminate($this->getRequest(), $response);
         }
+    }
+
+  /**
+   * Returns the factory method that can be used to retrieve the given kernel.
+   *
+   * @param string $kernel
+   *   The kernel to retrieve.
+   *
+   * @return callable
+   *   The factory method.
+   */
+    protected function getKernelFactory($kernel)
+    {
+        $factories = [
+          Kernels::DRUPAL => [DrushDrupalKernel::class, 'createFromRequest'],
+          Kernels::UPDATE => [DrushUpdateKernel::class, 'createFromRequest'],
+        ];
+        return $factories[$kernel];
     }
 }
