+++ src/Drupal/Commands/core/BatchCommands.php
@@ -13,6 +13,7 @@ class BatchCommands extends DrushCommands
      * @aliases batch-process
      * @param $batch_id The batch id that will be processed.
      * @hidden
+     * @kernel update
      */
     public function process($batch_id)
     {
